+++ src/Atata.WebDriverExtras/Extensions/StringBuilderExtensions.cs
@@ -3,7 +3,7 @@
 /// <summary>
 /// Provides a set of extension methods for <see cref="StringBuilder"/>.
 /// </summary>
-// TODO: v4. Remove StringBuilderExtensions.
+// TODO: v5. Remove StringBuilderExtensions.
 public static class StringBuilderExtensions
 {
     /// <summary>
@@ -11,6 +11,7 @@ public static class StringBuilderExtensions
     /// </summary>
     /// <param name="builder">The builder.</param>
     /// <returns>A reference to the same <see cref="StringBuilder"/> instance after the append operation has completed.</returns>
+    [Obsolete("Instead use Append(' ')")] // Obsolete since v4.0.0.
     public static StringBuilder AppendSpace(this StringBuilder builder)
     {
         Guard.ThrowIfNull(builder);