+++ src/Atata.WebDriverExtras/Extensions/IEnumerableExtensions.cs
@@ -1,9 +1,9 @@
 ﻿namespace Atata;
 
+// TODO: v5. Remove IEnumerableExtensions.
 public static class IEnumerableExtensions
 {
-    // TODO: v4. Make ToReadOnly method obsolete.
-    ////[Obsolete("Instead use constructor of ReadOnlyCollection<T> or use another collection type (array, for example).")] // Obsolete since v3.2.0.
+    [Obsolete("Instead use constructor of ReadOnlyCollection<T> or use another collection type (array, for example).")] // Obsolete since v4.0.0.
     public static ReadOnlyCollection<T> ToReadOnly<T>(this IEnumerable<T> source) =>
         new([.. source]);
 }