+++ src/Atata.WebDriverExtras/Extensions/StringBuilderExtensions.cs
@@ -11,7 +11,7 @@ public static class StringBuilderExtensions
     /// </summary>
     /// <param name="builder">The builder.</param>
     /// <returns>A reference to the same <see cref="StringBuilder"/> instance after the append operation has completed.</returns>
-    [Obsolete("Instead use Append(' ')")] // Obsolete since v4.0.0.
+    [Obsolete("Use Append(' ') instead.")] // Obsolete since v4.0.0.
     public static StringBuilder AppendSpace(this StringBuilder builder)
     {
         Guard.ThrowIfNull(builder);