+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Instead use `ElementNotFoundException.Create(...)`.
 - Make obsolete `ElementExceptionFactory.CreateForNotMissing` methods.
   Instead use `ElementNotMissingException.Create(...)`.
+- Make obsolete `StringBuilderExtensions.AppendSpace` method.
+  Instead use `Append(' ')`.
 
 ## [3.2.0] - 2025-04-08