+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Instead use constructor of `ReadOnlyCollection<T>` or use another collection type (array, for example).
 - Make obsolete `ExceptionFactory.CreateForTimeout` method.
   Use `TimeoutExceptionFactory.Create(...)` instead.
+- Make obsolete `ExceptionFactory.CreateForUnsupportedEnumValue` method.
 
 ### Removed