+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make obsolete `ExceptionFactory.CreateForTimeout` method.
   Use `TimeoutExceptionFactory.Create(...)` instead.
 - Make obsolete `ExceptionFactory.CreateForUnsupportedEnumValue` method.
+- Change `RetrySettings.DefaultInterval` value from 500 to 200 milliseconds.
 
 ### Removed