+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Upgrade Selenium.WebDriver package reference to v4.36.0.
+ 
 ## [3.2.0] - 2025-04-08
 
 ### Changed