+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Upgrade Selenium.WebDriver package reference to v4.36.0.
+- Add handling of `OpenQA.Selenium.UnknownErrorException` to `StaleSafely.Execute` methods.
  
 ## [3.2.0] - 2025-04-08