+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.3.0] - 2025-10-13
+
 ### Changed
 
 - Upgrade Selenium.WebDriver package reference to v4.36.0.
@@ -25,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Upgrade Selenium.WebDriver package reference to v4.28.0 (#61).
 
-[Unreleased]: https://github.com/atata-framework/atata-webdriverextras/compare/v3.2.0...HEAD
+[Unreleased]: https://github.com/atata-framework/atata-webdriverextras/compare/v3.3.0...HEAD
+[3.3.0]: https://github.com/atata-framework/atata-webdriverextras/compare/v3.2.0...v3.3.0
 [3.2.0]: https://github.com/atata-framework/atata-webdriverextras/compare/v3.1.0...v3.2.0
 [3.1.0]: https://github.com/atata-framework/atata-webdriverextras/compare/v3.0.0...v3.1.0